Added: Key data type for a better key handling
This commit introduces a new data type `Key` and its `KeyKind` companion type. The Key instance will hold a key information for the client to communicate with the backend.

We have 2 `KeyKind` at the moment,
1. Application Key: this key type is used when communicating with the `API` backend.
2. Public Key: this key type is used when communicating with the `Valut` backend.
Those 2 KeyKind are declared as empty enums and be used as a phantom type on `Key`