Refactor Key type and retrieve public key support (#15)
* Properly encode public key for the vault endpoints

* Append `:X` on both public and secret key

* Added: Key data type for a better key handling

This commit introduces a new data type `Key` and its `KeyKind` companion type. The Key instance will hold a key information for the client to communicate with the backend.

We have 2 `KeyKind` at the moment,
1. Application Key: this key type is used when communicating with the `API` backend.
2. Public Key: this key type is used when communicating with the `Valut` backend.
Those 2 KeyKind are declared as empty enums and be used as a phantom type on `Key`

* Adopting the `Key` data type in Omise Swift

Since we introduced the new `Key` data type. This commit migrated our Omise Swift code base to use `Key` data instead of a plain String.

* Added: PublicKeyEndpoint